requirements.txt
discord.py
apscheduler
pytz
certifi
